chore: Renamed Core/Kernel.mts to Core/kernel.service.mts.
