+++ packages/architectura/src/Core/kernel.service.mts
